+++ src/main/java/com/kbindiedev/verse/input/mouse/MouseButtons.java
@@ -1,0 +1,10 @@
+package com.kbindiedev.verse.input.mouse;
+
+/** A list of all mouse buttons */
+public class MouseButtons {
+
+    public static final int LEFT = 0;
+    public static final int RIGHT = 1;
+    public static final int MIDDLE = 2;
+
+}