+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Keyboard input handling system (event dispatch and querying functionality)
-- Keys list
+- Mouse input handling system (event dispatch and querying functionality)
+- Keys and MouseButtons list
 
 ---
 ## 0.0.1-indev-1 - 2021-11-28 - The init commit